LayoutTests/imported/w3c/web-platform-tests/css/css-tables/tentative/baseline-td-expected.txt

<TD> baseline tests

All the baseline values. Many text values map to plain baseline.
Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp
All the baseline values. Same as test before, but use display:table/table-cell. Default td alignment is baseline, not middle.
Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp	Xp
Baselines with %ge children
FF/Edge fails here
inline table

Baseline and td height If td is baseline aligned, and has a height, how tall should the td be?
Legacy/FF apply height to below baseline?. With baseline alignment, this makes the cell taller.
base	
Baseline of rowspanned cell rowspanned td's should contribute to baseline, but not to the row height.
Xp	Xp
Xp
Xp
Xp

PASS table, .display-table 1
PASS table, .display-table 2
FAIL table, .display-table 3 assert_equals:
<table style="height:100px;position:relative;display:inline-table">
    <tbody><tr><td style="vertical-align:baseline">
      <div style="width:100px;height:50%;background:rgba(0,120,0, 0.5)" data-offset-y="30"></div>
    </td>
     <td style="vertical-align:baseline">
      <div style="width:100px;height:80%;background:rgba(0,120,0, 0.5)" data-offset-y="0"></div>
    </td>
  </tr></tbody></table>
offsetTop expected 30 but got 0
PASS table, .display-table 4
PASS table, .display-table 5
